+++ pkg/app/app.go
@@ -4,8 +4,11 @@ import (
 	"encoding/json"
 	"fmt"
 	"io"
+	"strconv"
 
+	"github.com/giantswarm/apiextensions/v3/pkg/annotation"
 	applicationv1alpha1 "github.com/giantswarm/apiextensions/v3/pkg/apis/application/v1alpha1"
+	"github.com/giantswarm/apiextensions/v3/pkg/label"
 	"github.com/giantswarm/microerror"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"sigs.k8s.io/yaml"
@@ -16,6 +19,7 @@ type Config struct {
 	AppName             string
 	AppNamespace        string
 	AppVersion          string
+	ConfigMajorVersion  int
 	DisableForceUpgrade bool
 	Name                string
 	UserConfigMapName   string
@@ -26,18 +30,18 @@ type Config struct {
 //
 // AppCatalog is the name of the app catalog where the app stored.
 func NewCR(c Config) *applicationv1alpha1.App {
-	var annotations map[string]string
+	annotations := map[string]string{}
 	{
+		if c.ConfigMajorVersion > 0 {
+			annotations[annotation.ConfigMajorVersion] = strconv.Itoa(c.ConfigMajorVersion)
+		}
 		if !c.DisableForceUpgrade {
-			annotations = map[string]string{
-				"chart-operator.giantswarm.io/force-helm-upgrade": "true",
-			}
+			annotations["chart-operator.giantswarm.io/force-helm-upgrade"] = "true"
 		}
 	}
 
 	var userConfig applicationv1alpha1.AppSpecUserConfig
 	if c.UserConfigMapName != "" {
-
 		userConfig.ConfigMap = applicationv1alpha1.AppSpecUserConfigConfigMap{
 			Name:      c.UserConfigMapName,
 			Namespace: "giantswarm",
@@ -57,9 +61,10 @@ func NewCR(c Config) *applicationv1alpha1.App {
 			Namespace:   "giantswarm",
 			Annotations: annotations,
 			Labels: map[string]string{
-				// Version is set to 0.0.0 for all control plane app CRs so
-				// they are reconciled by app-operator-unique.
-				"app-operator.giantswarm.io/version": "0.0.0",
+				// Version 0.0.0 means this is reconciled by
+				// unique operator.
+				label.AppOperatorVersion:      "0.0.0",
+				label.ConfigControllerVersion: "0.0.0",
 			},
 		},
 		Spec: applicationv1alpha1.AppSpec{